requirements.txt
flask
flask-sqlalchemy
flask-security-too
celery
redis
bcrypt
matplotlib
email-validator
secure-smtplib